requirements.txt
python-vlc
kivy
